requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
ultralytics>=8.0.0
torch>=1.9.0
python-poker>=0.30.0
